Fixed more Sanitizer issues

+++ Emulator/Agnus/Blitter/BlitterTypes.h
@@ -26,10 +26,10 @@ typedef struct
     u16 ash;
     u16 bsh;
     u16 minterm;
-    u16 bltapt;
-    u16 bltbpt;
-    u16 bltcpt;
-    u16 bltdpt;
+    u32 bltapt;
+    u32 bltbpt;
+    u32 bltcpt;
+    u32 bltdpt;
     u16 bltafwm;
     u16 bltalwm;
     i16 bltamod;
